Combined trade commands to be a single command - added documentation
