+++ cmd/trade/trade.go
@@ -1,24 +1,194 @@
 package trade
 
 import (
-	
+	"fmt"
+	"math/rand"
+	"strings"
+	"time"
+
+	"github.com/AnnaGD/go-eth-trade-bot/cmd/constants"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/spf13/cobra"
 )
 
-// Trade command
+var (
 
+	// General trade parameters
+	tokenIn     string
+	tokenOut    string
+	amount      string
+	slippage    float64
+	deadlineMin uint
+
+	// Pool specific parameters
+	targetPool    string
+	imbalanceMode bool
+)
+
+// Execute command for trading
 var TradeCmd = &cobra.Command{
-	Use:   "trade",
-	Short: "Execute trades on Uniswap V2",
-	Long: `The trade command allows you to execute trades on Uniswap V2 liquidity pools.
-It supports swapping tokens to deliberately create imbalances in pools or regular trading.`,
+	Use:   "execute",
+	Short: "Execute a trade on Uniswap V2",
+	Long: `Execute a token swap on Uniswap V2 liquidity pools. This command allows you to swap tokens or create imbalances for testing.`,
+	Run: func(cmd *cobra.Command, args []string) {
+
+		// Get persistent flags
+		rpcURL, _ := cmd.Flags().GetString("rpc-url")
+		wallet, _ := cmd.Flags().GetString("wallet")
+
+		// keystoreFile, _ := cmd.Flags().GetString("keystore-file")
+		gasPrice, _ := cmd.Flags().GetString("gas-price")
+		gasLimit, _ := cmd.Flags().GetUint64("gas-limit")
+
+		// Connect to Network -- on this case Ethereum
+		client, err := ethclient.Dial(rpcURL)
+		if err != nil {
+			fmt.Printf("Error connecting to Ethereum: %v\n", err)
+			return
+		}
+
+		// Create a random number generator with its own source
+		rng := rand.New(rand.NewSource(time.Now().UnixNano()))
+
+		if imbalanceMode {
+			// Create deliberate imbalances for testing
+			fmt.Println("🔄 Starting imbalance trade...")
+
+			// If no specific pool, choose random pool
+			if targetPool == "" {
+
+				// Choose random pool
+				pools := make([]string, 0, len(constants.UniV2Pools))
+				for pool := range constants.UniV2Pools {
+					pools = append(pools, pool)
+				}
+				// Use the local RNG instead of global rand
+				targetPool = pools[rng.Intn(len(pools))]
+			}
+
+			// Get pool address
+			poolAddress, exists := constants.UniV2Pools[targetPool]
+			if !exists {
+				fmt.Printf("❌ Pool %s not found\n", targetPool)
+				return
+			}
+
+			fmt.Printf("🎯 Target Pool: %s (%s)\n", targetPool, poolAddress)
+
+			// Get token names from pool name
+			tokens := parsePoolTokens(targetPool)
+
+			// Choose which tokens to sell(randomly)
+			tokenToSell := tokens[rng.Intn(2)]
+			tokenToBuy := tokens[1-rng.Intn(2)] // The other token
+
+			fmt.Printf("💱 Creating imbalance by selling %s to buy %s\n", tokenToSell, tokenToBuy)
+			fmt.Printf("💰 Amount: %s\n", amount)
+
+			// Execute the trade
+			fmt.Println("Executing imbalance trade...")
+
+			// TODO: Connect to the pool contract and execute the swap
+			// This would be the same function that would be used for regular trading
+			// For demo, we'll simulate this
+			executeSwap(client, common.HexToAddress(poolAddress), tokenToSell, tokenToBuy, amount)
+
+			fmt.Println("✅ Trade complete! Pool is now imbalanced.")
+			fmt.Println("Arbitrage opportunity created for testing.")
+		} else {
+
+			// Regular Trading Mode: Normal token swapping
+			fmt.Println("🔄 Executing standard trade...")
+			fmt.Printf("  Token In: %s\n", tokenIn)
+			fmt.Printf("  Token Out: %s\n", tokenOut)
+			fmt.Printf("  Amount: %s\n", amount)
+			fmt.Printf("  Slippage: %.2f%%\n", slippage)
+			fmt.Printf("  Deadline: %d minutes\n", deadlineMin)
+			fmt.Printf("  RPC URL: %s\n", rpcURL)
+			fmt.Printf("  Wallet: %s\n", wallet)
+			fmt.Printf("  Gas Price: %s\n", gasPrice)
+			fmt.Printf("  Gas Limit: %d\n", gasLimit)
+
+			// TODO: Implement the actual trade execution
+			// This would involve:
+			// 1. Resolving token addresses
+			// 2. Setting up the transaction
+			// 3. Calculating minimum output with slippage
+			// 4. Executing the swap
+			fmt.Println("⚠️ Standard trade execution not fully implemented yet")
+
+		}
+	},
+}
+
+func parsePoolTokens(poolName string) []string {
+	tokens := []string{"TokenA", "TokenB"}
+
+	// Extract token names from format like "eEUR_eAUD_Pool"
+	if len(poolName) > 10 {
+		parts := strings.Split(poolName, "-")
+		if len(parts) >= 2 {
+			tokens[0] = parts[0]
+			tokens[1] = parts[1]
+		}
+	}
+
+	return tokens
+}
+
+// executeSwap is a placeholder for the actual swap execution
+func executeSwap(client *ethclient.Client, poolAddress common.Address, tokenIn, tokenOut, amount string) {
+	// TODO: Replace with actual swap logic
+	// This would:
+	// 1. Connect to the Uniswap V2 Router contract
+	// 2. Call the appropriate swap function
+	// 3. Handle transaction signing and broadcasting
+
+	// For demo, we'll just simulate a delay
+	fmt.Println("Sending transaction...")
+	time.Sleep(2 * time.Second)
+
+	// Simulate transaction success
+	fmt.Printf("Swapped %s %s for %s\n", amount, tokenIn, tokenOut)
 }
 
 func init() {
-	// Persistant flags for all trade subcommands
-	TradeCmd.PersistentFlags().StringP("rpc-url", "r", "https://eth-goerli.g.alchemy.com/v2/demo", "Ethereum RPC URL")
-	TradeCmd.PersistentFlags().StringP("wallet", "w", "", "Wallet address to use for trades")
-	TradeCmd.PersistentFlags().StringP("keystore-file", "k", "", "Path to keystore file")
-	TradeCmd.PersistentFlags().String("gas-price", "auto", "Gas price in Gwei or 'auto'")
-	TradeCmd.PersistentFlags().Uint64("gas-limit", 250000, "Gas limit for transactions")
+	// Add general trading flags
+	TradeCmd.Flags().StringVar(&tokenIn, "token-in", "ETH", "Input token symbol or address")
+	TradeCmd.Flags().StringVar(&tokenOut, "token-out", "", "Output token symbol or address")
+	TradeCmd.Flags().StringVar(&amount, "amount", "1.0", "Amount of input token to swap")
+	TradeCmd.Flags().Float64Var(&slippage, "slippage", 0.5, "Slippage tolerance percentage")
+	TradeCmd.Flags().UintVar(&deadlineMin, "deadline", 20, "Transaction deadline in minutes")
+
+	// Add pool-specific flags
+	TradeCmd.Flags().StringVar(&targetPool, "pool", "", "Target pool for trade")
+	TradeCmd.Flags().BoolVar(&imbalanceMode, "imbalance", false, "Create imbalance for testing arbitrage")
+
+	// Mark required flags for standard trading mode
+	// These are only checked when imbalanceMode is false
+	// TradeCmd.MarkFlagRequired("token-out")
+	// TradeCmd.MarkFlagRequired("amount")
+
+	// Add this command to the parent trade command
+	TradeCmd.AddCommand(TradeCmd)
 }
+
+/*
+The original execute.go was focused on general token swapping with parameters like:
+
+Input and output tokens
+Amount
+Slippage and deadline
+Wallet and keystore details
+
+The newer version introduces:
+
+Pool-specific trading through targetPool
+An imbalanceMode flag specifically for creating imbalances
+Parsing pool names to identify tokens
+Random token selection for creating imbalances
+
+The newer version is designed specifically for testing the arbitrage bot by deliberately creating imbalances in pools, while the original version was for general trading.
+
+*/
