+++ cmd/utils/pool_reader.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"math/big"
-
 	"github.com/ethereum/go-ethereum/ethclient"
 )
